+++ qase-javascript-commons/src/env/env-validation-schema.ts
@@ -24,12 +24,12 @@ export const envValidationSchema: JSONSchemaType<EnvType> = {
   properties: {
     [EnvEnum.mode]: {
       type: 'string',
-      enum: [ModeEnum.report, ModeEnum.testops, ModeEnum.off],
+      enum: [ModeEnum.report, ModeEnum.testops, ModeEnum.testops_multi, ModeEnum.off],
       nullable: true,
     },
     [EnvEnum.fallback]: {
       type: 'string',
-      enum: [ModeEnum.report, ModeEnum.testops, ModeEnum.off],
+      enum: [ModeEnum.report, ModeEnum.testops, ModeEnum.testops_multi, ModeEnum.off],
       nullable: true,
     },
     [EnvEnum.debug]: {